code/backend/requirements.txt
turicreate==6.4
fastapi==0.104.0
pyrebase4==4.7.1
uvicorn==0.24.0
pandas==2.0.3
numpy==1.24.3
